:books: [Docs]: Update AddnameSelectList.tsx
- Update common component of AddnameSelectList.tsx
- add file of style.ts in AddnameSelectList directory
