+++ src/components/index.ts
@@ -1,17 +1,17 @@
+export { default as LoadingComponent } from './Loadings/LoadingComponent';
+export { default as LoadingItem } from './Loadings/LoadingItem';
+export { default as LoadingLogo } from './Loadings/LoadingLogo';
+export { default as AddnameSelectList } from './AddnameSelectList';
 export { default as Background } from './Background';
 export { default as Containers } from './Containers';
 export { default as ErrorMessage } from './ErrorMessage';
 export { default as InputTitles } from './InputTitles';
-export { default as LoadingComponent } from './Loadings/LoadingComponent';
-export { default as LoadingItem } from './Loadings/LoadingItem';
-export { default as LoadingLogo } from './Loadings/LoadingLogo';
 export { default as NotFoundPage } from './NotFoundPage';
 export { default as Overlay } from './Overlay';
 export { default as ThemeButton } from './ThemeButton';
 export { default as DeleteConfirm } from './DeleteConfirm';
 export { default as KeyboardBox } from './KeyboardBox';
 export { default as DateComponent } from './DateComponent';
-export { default as AddnameSelectList } from './AddnameSelectList';
 export { default as Pagination } from './Pagination';
 export { default as PageRouterTitle } from './PageRouterTitle';
 export { default as DateKind } from './DateKind';
